data_pipeline/requirements.txt
# data_pipeline/requirements.txt
requests
psycopg2-binary